+++ components/cvi_mmf_sdk_cv180xx/include/cvi_middleware/cvi_mw_isp/isp_defines.h
@@ -326,6 +326,8 @@ CHECK_VALID_ARRAY_3D(_struct, stAuto._param, _idx1, _idx2, ISP_AUTO_ISO_STRENGTH
 #define ISP_PACKED ATTR((packed))
 #define ISP_ALIGNED(x) ATTR((aligned(x)))
 
+#define ENABLE_FE_WBG_UPDATE    0   // for rgbmap
+
 #ifdef __cplusplus
 #if __cplusplus
 }