+++ HISTORY.md
@@ -1,3 +1,8 @@
+update sdk g181b5742-819:
+1. update isp libs
+2. fix usb host drv bugs, ex. msc & RNDIS
+3. add esp32 support
+4. others components update
 
 update commit sdk_g8e6bcd58-756